% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStatR-package.R
\docType{package}
\name{fitStatR}
\alias{fitStatR}
\alias{fitStatR-package}
\title{fitStatR}
\description{
The fitStatR package calculates fit statistics for forecasting models.
}
\author{
JB Duck-Mayr
}
