% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{Calculate Mean Absolute Percentage Error (MAPE)}
\usage{
mape(y, P, ...)
}
\arguments{
\item{y}{A numeric vector of length n (the number of observations) giving
the observed outcomes}

\item{P}{A numeric matrix of n nows and m (the number of forecasting models)
columns, where each entry i,j is the predicted value of observation i
from forecasting model j}
}
\value{
A numeric vector of length n whose values are the MAPE for each
  forecasting model used to construct P
}
\description{
\code{mape} calculates MAPE for statistical models.
}
