% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meape.R
\name{meape}
\alias{meape}
\title{Calculate Median Absolute Percentage Error (MEAPE)}
\usage{
meape(y, P, ...)
}
\arguments{
\item{y}{A numeric vector of length n (the number of observations) giving
the observed outcomes}

\item{P}{A numeric matrix of n nows and m (the number of forecasting models)
columns, where each entry i,j is the predicted value of observation i
from forecasting model j}

\item{...}{Other arguments that will not be used by this function}
}
\value{
A numeric vector of length n whose values are the MEAPE for each
  forecasting model used to construct P
}
\description{
\code{meape} calculates MEAPE for statistical models.
}
