% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrae.R
\name{mrae}
\alias{mrae}
\title{Calculate Median Relative Absolute Error (MRAE)}
\usage{
mrae(y, P, r)
}
\arguments{
\item{y}{A numeric vector of length n (the number of observations) giving
the observed outcomes}

\item{P}{A numeric matrix of n nows and m (the number of forecasting models)
columns, where each entry i,j is the predicted value of observation i
from forecasting model j}

\item{r}{A numeric vector of naive forecasts}
}
\value{
A numeric vector of length n whose values are the MRAE for each
  forecasting model used to construct P
}
\description{
\code{mrae} calculates MRAE for statistical models.
}
