% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculate Root Mean Squared Error (RMSE)}
\usage{
rmse(y, P, ...)
}
\arguments{
\item{y}{A numeric vector of length n (the number of observations) giving
the observed outcomes}

\item{P}{A numeric matrix of n nows and m (the number of forecasting models)
columns, where each entry i,j is the predicted value of observation i
from forecasting model j}

\item{...}{Other arguments that will not be used by this function}
}
\value{
A numeric vector of length n whose values are the RMSE for each
  forecasting model used to construct P
}
\description{
\code{rmse} calculates RMSE for statistical models.
}
