% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStat.R
\name{fitStat}
\alias{fitStat}
\title{Calculate Fit Statistics for Forecasting Models}
\usage{
fitStat(y, P, r = NULL, stat = c("rmse", "mad", "rmsle", "mape", "meape"))
}
\arguments{
\item{y}{A numeric vector of length n (the number of observations) giving
the observed outcomes}

\item{P}{A numeric matrix of n nows and m (the number of forecasting models)
columns, where each entry i,j is the predicted value of observation i
from forecasting model j}

\item{r}{An optional numeric vector of naive forecasts; \code{r} must be
given if MRAE should be calculated}

\item{stat}{A character vector whose elements are the names of the fit
statistic functions to use; should be some combination of 'rmse', 'mad',
'rmsle', 'mape', 'meape', and 'mrae'. By default all applicable fit
statistic functions are used.}
}
\value{
A numeric matrix with m rows where each i,j entry is the relevant
  fit statistic for forecasting model i
}
\description{
\code{fitStat} calculates various fit statistics for statistical models.
}
\examples{
y <- 0:2
P <- matrix(c(1:3, 3:1, rep(2, 3)), nrow=3)
r <- rnorm(3)
fitStat(y, P) # warning about zero value(s) in y
fitStat(y, P, r) # warning about zero value(s) in y
y <- 1:3
fitStat(y, P)
fitStat(y, P, r)
fitStat(y, P, stat='rmse') # if you only want one of the fit statistics
fitStat(y, P, stat=c('rmse', 'rmsle')) # if you want some but not all
y <- c(NA, 2, 3)
fitStat(y, P) # warning about NA value(s) in y
fitStat(y, P, r) # warning about NA value(s) in y
y <- c(0, NA, 2)
fitStat(y, P) # warning about zero and NA value(s) in y
fitStat(y, P, r) # warning about zero and NA value(s) in y

}
